+++ CHANGELOG.md
@@ -41,6 +41,7 @@ and this project adheres to
 
 ### Changed
 
+- Order `state` is now a computed property
 - Split address fullname field into first_name and last_name fields
 - Update CourseSerializer to bind order and enrollment related to the user
 - Use a ViewSet to create address api